# Proyectos Spring io/ProyectoVideo/demo/docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"MirtiGame","description":"Proyecto video","contact":{"name":"Martin Vazquez","email":"martinnvazquez1212@gmail.com"},"version":"v1.0"},"servers":[{"url":"http://localhost:8080","description":"Proyecto video"}],"paths":{"/players":{"post":{"tags":["player-controller"],"summary":"Crear un nuevo Player","description":"Retorna el Player creado con la id, si ya existe ese username o email retorna 400Ademas el email debe ser valido y la Password contener:8 caracteres, una minuscula, una mayus y una especial","operationId":"savePlayer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Player"}}},"required":true},"responses":{"400":{"description":"Username o Email ya existente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Operacion realizada con exito","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}}}}},"/players/{id}":{"get":{"tags":["player-controller"],"summary":"Traer un Player por ID","description":"Retorna un Player si existe su ID, sino retorna error 404","operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Operacion realizada con exito","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Revisar que la app este corriendo","description":"Si la app esta corriendo retornara pong","operationId":"pong","responses":{"200":{"description":"Operacion realizada con exito","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}}},"components":{"schemas":{"Player":{"required":["email","password","userName"],"type":"object","properties":{"id":{"title":"Player ID","type":"integer","description":"La ID del Player","format":"int64","example":1},"userName":{"title":"Player username","type":"string","description":"username del Player","example":"bananirou"},"password":{"title":"Player password","type":"string","description":"password del Player","example":"Elsalto01#"},"email":{"title":"Player email","type":"string","description":"email del Player","example":"delpapu@email.com"},"avatar":{"title":"Player avatar url","type":"string","description":"url del avatar del Player","nullable":true,"example":"https://localhost:8080/avatars/myUsername"},"lastLogin":{"title":"Player lastLogin","type":"string","description":"Ultimo logueo Player","format":"date-time","nullable":true}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}}}}}